# openapi.yaml
openapi: 3.1.0
info:
  title: Silicon Waveform API
  version: 0.2.0
  summary: Provider-independent text-to-speech and speech-to-text
  description: |
    Synchronous speech operations for Carbons, Silicons, and IAM applications.
    TTS automatic fallback defaults to off; STT retains automatic fallback. TTS output is normalized to
    MP3 and stored in the represented actor's Briefcase. OBO requests are
    accepted only when the configured IAM and Briefcase request-bound contracts
    are available. Test planes use deterministic speech with paired IAM and
    Briefcase SDK uploads, delegated reads, and current replay access checks.
    Inbound OBO speech is unavailable until IAM supports downstream delegation
    from the initiating request. Bearer speech is implemented.
servers:
  - url: https://waveform.teamofsilicons.com/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /tts:
    post:
      security:
        - bearerAuth: []
        - oboAccess: []
      tags: [Speech]
      operationId: synthesizeSpeech
      summary: Convert text to speech and store the MP3 in Briefcase
      description: |
        Attempts Gemini 3.1 Flash TTS Preview, ElevenLabs Multilingual v2,
        then OpenAI tts-1. Returns only after the normalized MP3 is stored and
        both its permanent and temporary Briefcase URLs are available. A
        completed idempotent retry re-authorizes the stored file and returns a
        freshly issued temporary URL without rerunning synthesis.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/AppId'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TTSRequest' }
      responses:
        '200':
          description: Stored audio result, with a currently authorized temporary URL
          headers:
            X-Request-ID: { $ref: '#/components/headers/RequestId' }
            Idempotency-Replayed: { $ref: '#/components/headers/IdempotencyReplayed' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TTSResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/SourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '502': { $ref: '#/components/responses/BadGateway' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '504': { $ref: '#/components/responses/GatewayTimeout' }
        default: { $ref: '#/components/responses/Error' }
...
  /stt:
    post:
      security:
        - bearerAuth: []
        - oboAccess: []
      tags: [Speech]
      operationId: transcribeSpeech
      summary: Transcribe an authorized Briefcase file
      description: |
        The source must be within the originating Waveform application namespace.
        Use /stt/source-target to prepare an owner-authorized copy when necessary.
        Re-authorizes the permanent Briefcase reference as the represented
        actor, then attempts Gemini 3.5 Transcribe, OpenAI gpt-transcribe, and
        Deepgram Nova-3 Multilingual. Waveform never fetches the supplied URL
        directly. A completed idempotent retry verifies current access to the
        exact source before returning its cached transcript.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/AppId'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/STTRequest' }
      responses:
        '200':
          description: Normalized transcript released after current source authorization
          headers:
            X-Request-ID: { $ref: '#/components/headers/RequestId' }
            Idempotency-Replayed: { $ref: '#/components/headers/IdempotencyReplayed' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/STTResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/SourceNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '502': { $ref: '#/components/responses/BadGateway' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '504': { $ref: '#/components/responses/GatewayTimeout' }
        default: { $ref: '#/components/responses/Error' }
  /capabilities:
    get:
      tags: [Capabilities]
      operationId: getCapabilities
      summary: Get the stable Waveform-level speech capabilities
      security: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Normalized service capabilities, independent of live provider health
          headers:
            X-Request-ID: { $ref: '#/components/headers/RequestId' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Capabilities' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: IAM access token
      description: Opaque IAM access token for a Carbon or Silicon.
    oboAccess:
      type: apiKey
      in: header
      name: X-IAM-OBO-Access-Proof
      description: |
        Short-lived IAM proof bound to Waveform, the represented actor,
        organization, and requested action. X-App-ID is also required.
  parameters:
    OrgId:
      name: X-Org-ID
      in: header
      required: true
      description: Organization selected for this operation; it must match verified IAM state.
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: '^[A-Za-z0-9_-]+$'
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Scoped to actor, organization, and operation for at least 24 hours
        (the default is 24 hours). Reusing a key with a different validated
        request during retention returns 409.
      schema:
        type: string
        minLength: 8
        maxLength: 255
        pattern: '^[!-~]+$'
    AppId:
      name: X-App-ID
      in: header
      required: false
      description: Required with X-IAM-OBO-Access-Proof and omitted for Bearer requests.
      schema:
        type: string
        minLength: 1
        maxLength: 80
        pattern: '^[A-Za-z0-9_-]+$'
    RequestId:
      name: X-Request-ID
      in: header
      required: false
      description: |
        Optional attempt correlation UUID. Invalid values are replaced by
        Waveform. A successful idempotent replay returns the stored canonical
        operation UUID instead.
      schema: { type: string, format: uuid }
...
  headers:
    RequestId:
      description: |
        Waveform request correlation ID. A successful speech replay uses the
        stored canonical operation UUID; replay-time errors use the current
        attempt UUID.
      schema: { type: string, format: uuid }
    IdempotencyReplayed:
      description: |
        True when stable completed-operation state was reused. Any TTS
        temporary delivery URL in the response was freshly authorized and is
        not loaded from the idempotency record.
      schema: { type: boolean }
    CacheControl:
      description: Prevents intermediaries from retaining private speech data or correlation IDs.
      schema: { type: string, const: no-store }
  responses:
    BadRequest:
      description: Malformed headers, credentials, URL, language hint, or JSON input
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
        Cache-Control: { $ref: '#/components/headers/CacheControl' }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: IAM credentials are missing, invalid, expired, or inactive
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
        Cache-Control: { $ref: '#/components/headers/CacheControl' }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: The represented actor cannot use this organization or resource
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
        Cache-Control: { $ref: '#/components/headers/CacheControl' }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    SourceNotFound:
      description: Briefcase reports the exact STT source or completed TTS replay file as missing
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
        Cache-Control: { $ref: '#/components/headers/CacheControl' }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: Idempotency key mismatch or another request currently owns the key
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
        Retry-After: { schema: { type: integer, minimum: 1 } }
        Cache-Control: { $ref: '#/components/headers/CacheControl' }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    PayloadTooLarge:
      description: Text, request body, source media, or generated audio exceeds a synchronous limit
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
        Cache-Control: { $ref: '#/components/headers/CacheControl' }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    TooManyRequests:
      description: Waveform admission control rejected the request
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
        Retry-After: { schema: { type: integer, minimum: 1 } }
        Cache-Control: { $ref: '#/components/headers/CacheControl' }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    BadGateway:
      description: Selected TTS provider failed with fallback off, or the enabled provider chain failed
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
        Cache-Control: { $ref: '#/components/headers/CacheControl' }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ServiceUnavailable:
      description: IAM, Briefcase, PostgreSQL, or a required contract capability is unavailable
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
        Retry-After: { schema: { type: integer, minimum: 1 } }
        Cache-Control: { $ref: '#/components/headers/CacheControl' }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    GatewayTimeout:
      description: The bounded synchronous operation exceeded its deadline
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
        Cache-Control: { $ref: '#/components/headers/CacheControl' }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Error:
      description: Normalized Waveform error
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestId' }
        Cache-Control: { $ref: '#/components/headers/CacheControl' }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    VoiceProfileRef:
      type: object
      required: [id, revision]
      properties:
        id: { type: string }
        revision: { type: integer, minimum: 1 }
...
    TTSRequest:
      type: object
      additionalProperties: false
      required: [text]
      properties:
        auto_fallback:
          type: boolean
          default: false
          description: Try the remaining provider order only when true. Nonempty provider_options require false.
        provider_options:
          $ref: '#/components/schemas/TTSProviderOptions'
        provider_keys:
          $ref: '#/components/schemas/TTSProviderKeys'
        voice_profile:
          type: [string, 'null']
          pattern: '^[a-z0-9-]{1,64}$'
          description: Optional profile ID from voice-profiles. Omitted/null uses the account default. Applies to all fallback providers and never changes the saved default.
        text:
          type: string
          minLength: 1
          maxLength: 4096
          description: Non-whitespace text. The limit counts Unicode scalar values.
        lang:
          type: string
          minLength: 2
          maxLength: 64
          description: Optional canonicalizable BCP 47 hint; passed only to Gemini TTS.
        provider_order:
          type: array
          minItems: 1
          maxItems: 3
          uniqueItems: true
          description: Preferred provider prefix; only its first resolved provider runs unless auto_fallback is true.
          items:
            type: string
            enum: [gemini, elevenlabs, openai]
    STTRequest:
      type: object
      additionalProperties: false
      required: [file_url]
      properties:
        provider_keys:
          $ref: '#/components/schemas/STTProviderKeys'
        file_url:
          type: string
          format: uri
          pattern: '^https://'
          description: Permanent authenticated URL at the configured Silicon Briefcase origin.
        language:
          type: string
          minLength: 2
          maxLength: 64
          description: Optional canonicalizable BCP 47 transcription hint.
        provider_order:
          type: array
          minItems: 1
          maxItems: 3
          uniqueItems: true
          description: Preferred provider prefix; omitted providers follow the account order.
          items:
            type: string
            enum: [gemini, openai, deepgram]
    TTSResult:
      type: object
      additionalProperties: false
      required: [request_id, file_url, temporary_url, media_type, provider, duration_ms]
      properties:
        voice_profile:
          oneOf:
            - $ref: '#/components/schemas/VoiceProfileRef'
            - { type: 'null' }
        request_id:
          type: string
          format: uuid
          description: Canonical operation UUID, preserved on completed replay.
        file_url: { type: string, format: uri, description: Permanent authenticated Briefcase URL }
        temporary_url:
          type: [string, 'null']
          format: uri
          description: Optional expiring delivery URL; null for published Briefcase one-shot uploads.
        media_type: { type: string, const: audio/mpeg }
        provider:
          type: string
          enum: [gemini, elevenlabs, openai]
          description: Provider that completed synthesis
        duration_ms: { type: integer, format: int64, minimum: 0 }
    STTResult:
      type: object
      additionalProperties: false
      required: [request_id, transcript, detected_language, provider, duration_ms]
      properties:
        request_id:
          type: string
          format: uuid
          description: Canonical operation UUID, preserved on completed replay.
        transcript: { type: string }
        detected_language: { type: [string, 'null'] }
        provider:
          type: string
          enum: [gemini, openai, deepgram]
          description: Provider that completed transcription
        duration_ms: { type: [integer, 'null'], format: int64, minimum: 0 }
...